+++ x-pack/plugins/enterprise_search/public/applications/shared/telemetry/send_telemetry.test.tsx
@@ -7,7 +7,7 @@
 import React from 'react';
 
 import { httpServiceMock } from 'src/core/public/mocks';
-import { mountWithKibanaContext } from '../../test_utils/helpers';
+import { mountWithKibanaContext } from '../../test_utils';
 import { sendTelemetry, SendAppSearchTelemetry } from './';
 
 describe('Shared Telemetry Helpers', () => {
