+++ x-pack/plugins/enterprise_search/public/applications/test_utils/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { mockKibanaContext } from './mock_kibana_context';
+export { mountWithKibanaContext } from './mount_with_context';
+
+// Note: mock_shallow_usecontext must be imported directly as a file
